NODE-1037: Move single validation methods to an object.
An implementation of `Validation` trait provides "high-level" methods that are composed of smaller ones. Also, please note, that `ProtoUtil.basicDeploy` will now generate signed deploy with correct headers.
