+++ casper/src/test/scala/io/casperlabs/casper/ValidationTest.scala
@@ -199,12 +199,12 @@ class ValidationTest
         block1 <- dagStorage
                    .lookupByIdUnsafe(1)
                    .map(_.changeSigAlgorithm(rsa))
-        _ <- Validation[Task].blockSignature(block0) shouldBeF false
+        _ <- Validation.blockSignature[Task](block0) shouldBeF false
         _ = log.warns.last
           .contains(s"signature algorithm '$unknownAlgorithm' is unsupported") should be(
           true
         )
-        _ <- Validation[Task].blockSignature(block1) shouldBeF false
+        _ <- Validation.blockSignature[Task](block1) shouldBeF false
         result = log.warns.last.contains(s"signature algorithm '$rsa' is unsupported") should be(
           true
         )
@@ -226,7 +226,7 @@ class ValidationTest
         block4       <- signedBlock(4).map(_.changeSig(invalidKey))
         block5       <- signedBlock(5).map(_.changeSig(block0.getSignature.sig)) //wrong sig
         blocks       = Vector(block0, block1, block2, block3, block4, block5)
-        _            <- blocks.existsM[Task](b => Validation[Task].blockSignature(b)) shouldBeF false
+        _            <- blocks.existsM[Task](b => Validation.blockSignature[Task](b)) shouldBeF false
         _            = log.warns.size should be(blocks.length)
         result       = log.warns.forall(_.contains("signature is invalid")) should be(true)
       } yield result
@@ -251,12 +251,12 @@ class ValidationTest
       sk,
       Ed25519
     )
-    Validation[Task].blockSignature(block) shouldBeF true
+    Validation.blockSignature[Task](block) shouldBeF true
   }
 
   "Deploy signature validation" should "return true for valid signatures" in withoutStorage {
     val deploy = sample(arbitrary[consensus.Deploy])
-    Validation[Task].deploySignature(deploy) shouldBeF true
+    Validation.deploySignature[Task](deploy) shouldBeF true
   }
 
   it should "return false if a key in an approval is empty" in withoutStorage {
@@ -265,7 +265,7 @@ class ValidationTest
     } yield d.withApprovals(d.approvals.map(x => x.withApproverPublicKey(ByteString.EMPTY)))
 
     val deploy = sample(genDeploy)
-    Validation[Task].deploySignature(deploy) shouldBeF false
+    Validation.deploySignature[Task](deploy) shouldBeF false
   }
 
   it should "return false for missing signatures" in withoutStorage {
@@ -274,7 +274,7 @@ class ValidationTest
     } yield d.withApprovals(Nil)
 
     val deploy = sample(genDeploy)
-    Validation[Task].deploySignature(deploy) shouldBeF false
+    Validation.deploySignature[Task](deploy) shouldBeF false
   }
 
   it should "return false for invalid signatures" in withoutStorage {
@@ -284,7 +284,7 @@ class ValidationTest
     } yield d.withApprovals(d.approvals.map(a => a.withSignature(a.getSignature.withSig(h))))
 
     val deploy = sample(genDeploy)
-    Validation[Task].deploySignature(deploy) shouldBeF false
+    Validation.deploySignature[Task](deploy) shouldBeF false
   }
 
   it should "return false if there are valid and invalid signatures mixed" in withoutStorage {
@@ -296,7 +296,7 @@ class ValidationTest
     )
 
     val deploy = sample(genDeploy)
-    Validation[Task].deploySignature(deploy) shouldBeF false
+    Validation.deploySignature[Task](deploy) shouldBeF false
   }
 
   "Deploy header validation" should "accept valid headers" in {
@@ -304,40 +304,40 @@ class ValidationTest
       ConsensusConfig(dagSize = 10, maxSessionCodeBytes = 5, maxPaymentCodeBytes = 5)
 
     forAll { (deploy: consensus.Deploy) =>
-      withoutStorage { Validation[Task].deployHeader(deploy, chainName) } shouldBe Nil
+      withoutStorage { Validation.deployHeader[Task](deploy, chainName) } shouldBe Nil
     }
   }
 
   it should "not accept too short time to live" in withoutStorage {
     val deploy = DeployOps.randomTooShortTTL()
-    Validation[Task].deployHeader(deploy, chainName) shouldBeF List(
+    Validation.deployHeader[Task](deploy, chainName) shouldBeF List(
       DeployHeaderError
-        .timeToLiveTooShort(deploy.deployHash, deploy.getHeader.ttlMillis, ValidationImpl.MIN_TTL)
+        .timeToLiveTooShort(deploy.deployHash, deploy.getHeader.ttlMillis, Validation.MIN_TTL)
     )
   }
 
   it should "not accept too long time to live" in withoutStorage {
     val deploy = DeployOps.randomTooLongTTL()
-    Validation[Task].deployHeader(deploy, chainName) shouldBeF List(
+    Validation.deployHeader[Task](deploy, chainName) shouldBeF List(
       DeployHeaderError
-        .timeToLiveTooLong(deploy.deployHash, deploy.getHeader.ttlMillis, ValidationImpl.MAX_TTL)
+        .timeToLiveTooLong(deploy.deployHash, deploy.getHeader.ttlMillis, Validation.MAX_TTL)
     )
   }
 
   it should "not accept too many dependencies" in withoutStorage {
     val deploy = DeployOps.randomTooManyDependencies()
-    Validation[Task].deployHeader(deploy, chainName) shouldBeF List(
+    Validation.deployHeader[Task](deploy, chainName) shouldBeF List(
       DeployHeaderError.tooManyDependencies(
         deploy.deployHash,
         deploy.getHeader.dependencies.size,
-        ValidationImpl.MAX_DEPENDENCIES
+        Validation.MAX_DEPENDENCIES
       )
     )
   }
 
   it should "not accept invalid dependencies" in withoutStorage {
     val deploy = DeployOps.randomInvalidDependency()
-    Validation[Task].deployHeader(deploy, chainName) shouldBeF List(
+    Validation.deployHeader[Task](deploy, chainName) shouldBeF List(
       DeployHeaderError
         .invalidDependency(deploy.deployHash, deploy.getHeader.dependencies.head)
     )
@@ -349,7 +349,7 @@ class ValidationTest
       arbitrary[consensus.Deploy].map(_.withChainName(s"never say $chainName"))
     }
 
-    Validation[Task].deployHeader(deploy, chainName) shouldBeF List(
+    Validation.deployHeader[Task](deploy, chainName) shouldBeF List(
       DeployHeaderError
         .invalidChainName(deploy.deployHash, deploy.getHeader.chainName, chainName)
     )
@@ -361,12 +361,12 @@ class ValidationTest
         _                       <- createChain[Task](1)
         block                   <- dagStorage.lookupByIdUnsafe(0)
         modifiedTimestampHeader = block.header.get.withTimestamp(99999999)
-        _ <- ValidationImpl[Task]
-              .timestamp(
+        _ <- Validation
+              .timestamp[Task](
                 block.withHeader(modifiedTimestampHeader)
               )
               .attempt shouldBeF Left(InvalidUnslashableBlock)
-        _      <- ValidationImpl[Task].timestamp(block).attempt shouldBeF Right(())
+        _      <- Validation.timestamp[Task](block).attempt shouldBeF Right(())
         _      = log.warns.size should be(1)
         result = log.warns.head.contains("block timestamp") should be(true)
       } yield result
@@ -378,12 +378,12 @@ class ValidationTest
         _                       <- createChain[Task](2)
         block                   <- dagStorage.lookupByIdUnsafe(1)
         modifiedTimestampHeader = block.header.get.withTimestamp(-1)
-        _ <- ValidationImpl[Task]
-              .timestamp(
+        _ <- Validation
+              .timestamp[Task](
                 block.withHeader(modifiedTimestampHeader)
               )
               .attempt shouldBeF Left(InvalidUnslashableBlock)
-        _      <- ValidationImpl[Task].timestamp(block).attempt shouldBeF Right(())
+        _      <- Validation.timestamp[Task](block).attempt shouldBeF Right(())
         _      = log.warns.size should be(1)
         result = log.warns.head.contains("block timestamp") should be(true)
       } yield result
@@ -402,12 +402,12 @@ class ValidationTest
             )
         block4                  <- dagStorage.lookupByIdUnsafe(4)
         modifiedTimestampHeader = block4.header.get.withTimestamp(genesis.getHeader.timestamp + 1)
-        _ <- ValidationImpl[Task]
-              .timestamp(
+        _ <- Validation
+              .timestamp[Task](
                 block4.withHeader(modifiedTimestampHeader)
               )
               .attempt shouldBeF Left(InvalidUnslashableBlock)
-        _      <- ValidationImpl[Task].timestamp(block4).attempt shouldBeF Right(())
+        _      <- Validation.timestamp[Task](block4).attempt shouldBeF Right(())
         _      = log.warns.size should be(1)
         result = log.warns.head.contains("block timestamp") should be(true)
       } yield result
@@ -419,12 +419,10 @@ class ValidationTest
         _     <- createChain[Task](1)
         block <- dagStorage.lookupByIdUnsafe(0)
         dag   <- dagStorage.getRepresentation
-        _ <- ValidationImpl[Task]
-              .blockRank(block.changeBlockNumber(1), dag)
-              .attempt shouldBeF Left(
+        _ <- Validation.blockRank[Task](block.changeBlockNumber(1), dag).attempt shouldBeF Left(
               InvalidBlockNumber
             )
-        _ <- ValidationImpl[Task].blockRank(block, dag) shouldBeF Unit
+        _ <- Validation.blockRank[Task](block, dag) shouldBeF Unit
         _ = log.warns.size should be(1)
         result = log.warns.head.contains("not zero, but block has no justifications") should be(
           true
@@ -439,15 +437,15 @@ class ValidationTest
       for {
         _   <- createChain[Task](n.toInt, bonds = List(Bond(validator, 1)), creator = validator)
         dag <- dagStorage.getRepresentation
-        _   <- dagStorage.lookupByIdUnsafe(0) >>= (b => ValidationImpl[Task].blockRank(b, dag))
-        _   <- dagStorage.lookupByIdUnsafe(1) >>= (b => ValidationImpl[Task].blockRank(b, dag))
-        _   <- dagStorage.lookupByIdUnsafe(2) >>= (b => ValidationImpl[Task].blockRank(b, dag))
-        _   <- dagStorage.lookupByIdUnsafe(3) >>= (b => ValidationImpl[Task].blockRank(b, dag))
-        _   <- dagStorage.lookupByIdUnsafe(4) >>= (b => ValidationImpl[Task].blockRank(b, dag))
-        _   <- dagStorage.lookupByIdUnsafe(5) >>= (b => ValidationImpl[Task].blockRank(b, dag))
+        _   <- dagStorage.lookupByIdUnsafe(0) >>= (b => Validation.blockRank[Task](b, dag))
+        _   <- dagStorage.lookupByIdUnsafe(1) >>= (b => Validation.blockRank[Task](b, dag))
+        _   <- dagStorage.lookupByIdUnsafe(2) >>= (b => Validation.blockRank[Task](b, dag))
+        _   <- dagStorage.lookupByIdUnsafe(3) >>= (b => Validation.blockRank[Task](b, dag))
+        _   <- dagStorage.lookupByIdUnsafe(4) >>= (b => Validation.blockRank[Task](b, dag))
+        _   <- dagStorage.lookupByIdUnsafe(5) >>= (b => Validation.blockRank[Task](b, dag))
         _ <- (0 until n).toList.forallM[Task] { i =>
               (dagStorage.lookupByIdUnsafe(i) >>= (
-                  b => ValidationImpl[Task].blockRank(b, dag)
+                  b => Validation.blockRank[Task](b, dag)
               )).map(_ => true)
             } shouldBeF true
         result = log.warns should be(Nil)
@@ -477,10 +475,8 @@ class ValidationTest
         b2  <- createBlockWithNumber(7)
         b3  <- createBlockWithNumber(8, Seq(b1, b2))
         dag <- dagStorage.getRepresentation
-        _   <- ValidationImpl[Task].blockRank(b3, dag) shouldBeF Unit
-        result <- ValidationImpl[Task]
-                   .blockRank(b3.changeBlockNumber(4), dag)
-                   .attempt shouldBeF Left(
+        _   <- Validation.blockRank[Task](b3, dag) shouldBeF Unit
+        result <- Validation.blockRank[Task](b3.changeBlockNumber(4), dag).attempt shouldBeF Left(
                    InvalidBlockNumber
                  )
       } yield result
@@ -496,15 +492,15 @@ class ValidationTest
           "Justification list of Genesis block should be empty."
         )
         dag <- dagStorage.getRepresentation
-        _ <- ValidationImpl[Task]
-              .sequenceNumber(
+        _ <- Validation
+              .sequenceNumber[Task](
                 block.withHeader(block.getHeader.withValidatorBlockSeqNum(1)),
                 dag
               )
               .attempt shouldBeF Left(
               InvalidSequenceNumber
             )
-        _ <- ValidationImpl[Task].sequenceNumber(
+        _ <- Validation.sequenceNumber[Task](
               block.withHeader(block.getHeader.withValidatorBlockSeqNum(0)),
               dag
             ) shouldBeF Unit
@@ -517,8 +513,8 @@ class ValidationTest
         _     <- createChainWithRoundRobinValidators[Task](2, 2)
         block <- dagStorage.lookupByIdUnsafe(1)
         dag   <- dagStorage.getRepresentation
-        _ <- ValidationImpl[Task]
-              .sequenceNumber(
+        _ <- Validation
+              .sequenceNumber[Task](
                 block.withHeader(block.getHeader.withValidatorBlockSeqNum(2)),
                 dag
               )
@@ -540,7 +536,7 @@ class ValidationTest
                 for {
                   block <- dagStorage.lookupByIdUnsafe(i)
                   dag   <- dagStorage.getRepresentation
-                  _ <- ValidationImpl[Task].sequenceNumber(
+                  _ <- Validation.sequenceNumber[Task](
                         block,
                         dag
                       )
@@ -559,7 +555,7 @@ class ValidationTest
         b1  <- createAndStoreBlockFull[Task](v2, List(b0), List(b0))
         b2  <- createAndStoreBlockFull[Task](v1, List(b1), List(b1))
         dag <- dagStorage.getRepresentation
-        _   <- ValidationImpl[Task].validatorPrevBlockHash(b2.getSummary, dag)
+        _   <- Validation.validatorPrevBlockHash[Task](b2.getSummary, dag)
       } yield ()
   }
   it should "pass if the hash is in the justifications" in withStorage {
@@ -570,7 +566,7 @@ class ValidationTest
         b0  <- createAndStoreBlockFull[Task](v1, List(g), Nil)
         b1  <- createAndStoreBlockFull[Task](v1, List(b0), List(b0))
         dag <- dagStorage.getRepresentation
-        _   <- ValidationImpl[Task].validatorPrevBlockHash(b1.getSummary, dag)
+        _   <- Validation.validatorPrevBlockHash[Task](b1.getSummary, dag)
       } yield ()
   }
   it should "fail if the hash belongs to somebody else" in withStorage {
@@ -587,7 +583,7 @@ class ValidationTest
                maybeValidatorPrevBlockHash = Some(b1.blockHash)
              )
         dag    <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task].validatorPrevBlockHash(b2.getSummary, dag).attempt
+        result <- Validation.validatorPrevBlockHash[Task](b2.getSummary, dag).attempt
       } yield {
         result shouldBe Left(ValidateErrorWrapper(InvalidPrevBlockHash))
       }
@@ -607,7 +603,7 @@ class ValidationTest
                maybeValidatorPrevBlockHash = Some(b1.blockHash)
              )
         dag    <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task].validatorPrevBlockHash(b3.getSummary, dag).attempt
+        result <- Validation.validatorPrevBlockHash[Task](b3.getSummary, dag).attempt
       } yield {
         result shouldBe Left(ValidateErrorWrapper(InvalidPrevBlockHash))
       }
@@ -626,7 +622,7 @@ class ValidationTest
                maybeValidatorBlockSeqNum = Some(1)
              )
         dag    <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task].validatorPrevBlockHash(b0.getSummary, dag).attempt
+        result <- Validation.validatorPrevBlockHash[Task](b0.getSummary, dag).attempt
       } yield {
         result shouldBe Left(ValidateErrorWrapper(InvalidPrevBlockHash))
       }
@@ -645,8 +641,8 @@ class ValidationTest
         invalidBlock <- dagStorage
                          .lookupByIdUnsafe(2)
                          .map(_.changeValidator(impostor))
-        _      <- Validation[Task].blockSender(validBlock) shouldBeF true
-        result <- Validation[Task].blockSender(invalidBlock) shouldBeF false
+        _      <- Validation.blockSender[Task](validBlock) shouldBeF true
+        result <- Validation.blockSender[Task](invalidBlock) shouldBeF false
       } yield result
   }
 
@@ -844,14 +840,12 @@ class ValidationTest
                 genesis,
                 dag
               )
-        _                 <- Validation[Task].bondsCache(genesis, genesisBonds) shouldBeF Unit
+        _                 <- Validation.bondsCache[Task](genesis, genesisBonds) shouldBeF Unit
         modifiedBonds     = Seq.empty[Bond]
         modifiedPostState = genesis.getHeader.getState.withBonds(modifiedBonds)
         modifiedHeader    = genesis.getHeader.withState(modifiedPostState)
         modifiedGenesis   = genesis.withHeader(modifiedHeader)
-        result <- Validation[Task]
-                   .bondsCache(modifiedGenesis, genesisBonds)
-                   .attempt shouldBeF Left(
+        result <- Validation.bondsCache[Task](modifiedGenesis, genesisBonds).attempt shouldBeF Left(
                    InvalidBondsCache
                  )
       } yield result
@@ -865,27 +859,27 @@ class ValidationTest
       val genesis                               = ProtoUtil.signBlock(block, sk, Ed25519)
 
       for {
-        _ <- Validation[Task].formatOfFields(genesis) shouldBeF true
-        _ <- Validation[Task].formatOfFields(genesis.withBlockHash(ByteString.EMPTY)) shouldBeF false
-        _ <- Validation[Task].formatOfFields(genesis.clearHeader) shouldBeF false
-        _ <- Validation[Task].formatOfFields(genesis.clearBody) shouldBeF true // Body is only checked in `Validate.blockFull`
-        _ <- Validation[Task].formatOfFields(
+        _ <- Validation.formatOfFields[Task](genesis) shouldBeF true
+        _ <- Validation.formatOfFields[Task](genesis.withBlockHash(ByteString.EMPTY)) shouldBeF false
+        _ <- Validation.formatOfFields[Task](genesis.clearHeader) shouldBeF false
+        _ <- Validation.formatOfFields[Task](genesis.clearBody) shouldBeF true // Body is only checked in `Validate.blockFull`
+        _ <- Validation.formatOfFields[Task](
               genesis.withSignature(genesis.getSignature.withSig(ByteString.EMPTY))
             ) shouldBeF false
-        _ <- Validation[Task].formatOfFields(
+        _ <- Validation.formatOfFields[Task](
               genesis.withSignature(genesis.getSignature.withSigAlgorithm(""))
             ) shouldBeF false
-        _ <- Validation[Task].formatOfFields(
+        _ <- Validation.formatOfFields[Task](
               genesis.withHeader(genesis.getHeader.withChainName(""))
             ) shouldBeF false
-        _ <- Validation[Task].formatOfFields(genesis.withHeader(genesis.getHeader.clearState)) shouldBeF false
-        _ <- Validation[Task].formatOfFields(
+        _ <- Validation.formatOfFields[Task](genesis.withHeader(genesis.getHeader.clearState)) shouldBeF false
+        _ <- Validation.formatOfFields[Task](
               genesis.withHeader(
                 genesis.getHeader
                   .withState(genesis.getHeader.getState.withPostStateHash(ByteString.EMPTY))
               )
             ) shouldBeF false
-        _ <- Validation[Task].formatOfFields(
+        _ <- Validation.formatOfFields[Task](
               genesis.withHeader(genesis.getHeader.withBodyHash(ByteString.EMPTY))
             ) shouldBeF false
       } yield ()
@@ -898,12 +892,12 @@ class ValidationTest
     } yield d.withDeployHash(h)
 
     val deploy = sample(genDeploy)
-    Validation[Task].deployHash(deploy) shouldBeF false
+    Validation.deployHash[Task](deploy) shouldBeF false
   }
 
   it should "return true for valid hashes" in {
     val deploy = sample(arbitrary[consensus.Deploy])
-    Validation[Task].deployHash(deploy) shouldBeF true
+    Validation.deployHash[Task](deploy) shouldBeF true
   }
 
   "Processed deploy validation" should "fail a block with a deploy having an invalid hash" in withStorage {
@@ -920,7 +914,7 @@ class ValidationTest
         )
       }
       for {
-        result <- ValidationImpl[Task].deployHashes(block).attempt
+        result <- Validation.deployHashes[Task](block).attempt
         _      = result shouldBe Left(ValidateErrorWrapper(InvalidDeployHash))
       } yield ()
   }
@@ -937,7 +931,7 @@ class ValidationTest
       )
     }
     for {
-      result <- ValidationImpl[Task].deploySignatures(block).attempt
+      result <- Validation.deploySignatures[Task](block).attempt
       _      = result shouldBe Left(ValidateErrorWrapper(InvalidDeploySignature))
     } yield ()
   }
@@ -964,7 +958,7 @@ class ValidationTest
       )
     }
     for {
-      result <- ValidationImpl[Task].deploySignatures(block).attempt
+      result <- Validation.deploySignatures[Task](block).attempt
       _      = result shouldBe Left(ValidateErrorWrapper(InvalidDeploySignature))
     } yield ()
   }
@@ -982,8 +976,8 @@ class ValidationTest
       }
       for {
         dag <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task]
-                   .deployHeaders(
+        result <- Validation
+                   .deployHeaders[Task](
                      block,
                      dag,
                      chainName = "no country for old men"
@@ -1013,7 +1007,7 @@ class ValidationTest
       }
       for {
         dag <- dagStorage.getRepresentation
-        _   <- ValidationImpl[Task].deployHeaders(block, dag, chainName = "area 51")
+        _   <- Validation.deployHeaders[Task](block, dag, chainName = "area 51")
       } yield ()
   }
 
@@ -1023,9 +1017,9 @@ class ValidationTest
       HashSetCasperTest.createGenesis(Map(pk -> 1))
     val signedBlock = ProtoUtil.signBlock(block, sk, Ed25519)
     for {
-      _ <- ValidationImpl[Task].blockHash(signedBlock) shouldBeF Unit
-      result <- ValidationImpl[Task]
-                 .blockHash(
+      _ <- Validation.blockHash[Task](signedBlock) shouldBeF Unit
+      result <- Validation
+                 .blockHash[Task](
                    signedBlock.withBlockHash(ByteString.copyFromUtf8("123"))
                  )
                  .attempt shouldBeF Left(InvalidBlockHash)
@@ -1039,9 +1033,9 @@ class ValidationTest
         HashSetCasperTest.createGenesis(Map(pk -> 1))
       val signedBlock = ProtoUtil.signBlock(block, sk, Ed25519)
       for {
-        _ <- ValidationImpl[Task].deployCount(signedBlock) shouldBeF Unit
-        result <- ValidationImpl[Task]
-                   .deployCount(
+        _ <- Validation.deployCount[Task](signedBlock) shouldBeF Unit
+        result <- Validation
+                   .deployCount[Task](
                      signedBlock.withHeader(signedBlock.header.get.withDeployCount(100))
                    )
                    .attempt shouldBeF Left(InvalidDeployCount)
@@ -1062,11 +1056,11 @@ class ValidationTest
     val signedBlock = ProtoUtil.signBlock(block, sk, Ed25519)
 
     for {
-      _ <- Validation[Task].version(
+      _ <- Validation.version[Task](
             signedBlock,
             missingProtocolVersionForBlock
           ) shouldBeF false
-      result <- Validation[Task].version(
+      result <- Validation.version[Task](
                  signedBlock,
                  protocolVersionForGenesisBlock
                ) shouldBeF true
@@ -1111,7 +1105,7 @@ class ValidationTest
       for {
         block  <- createBlock[Task](Seq.empty, deploys = deploysWithCost)
         dag    <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task].deployHeaders(block, dag, chainName).attempt
+        result <- Validation.deployHeaders[Task](block, dag, chainName).attempt
       } yield result shouldBe Left(ValidateErrorWrapper(InvalidDeployHeader))
   }
 
@@ -1140,7 +1134,7 @@ class ValidationTest
                   .map(_.changeTimestamp(blockTimestamp))
         _      <- blockStorage.put(block.blockHash, block, Seq.empty)
         dag    <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task].deployHeaders(block, dag, chainName).attempt
+        result <- Validation.deployHeaders[Task](block, dag, chainName).attempt
       } yield result shouldBe Left(ValidateErrorWrapper(DeployFromFuture))
   }
 
@@ -1153,7 +1147,7 @@ class ValidationTest
                   .map(_.changeTimestamp(blockTimestamp))
         _      <- blockStorage.put(block.blockHash, block, Seq.empty)
         dag    <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task].deployHeaders(block, dag, chainName).attempt
+        result <- Validation.deployHeaders[Task](block, dag, chainName).attempt
       } yield result shouldBe Left(ValidateErrorWrapper(DeployExpired))
   }
 
@@ -1167,7 +1161,7 @@ class ValidationTest
                   .map(_.changeTimestamp(blockTimestamp))
         _      <- blockStorage.put(block.blockHash, block, Seq.empty)
         dag    <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task].deployHeaders(block, dag, chainName).attempt
+        result <- Validation.deployHeaders[Task](block, dag, chainName).attempt
       } yield result shouldBe Left(ValidateErrorWrapper(DeployDependencyNotMet))
   }
 
@@ -1196,7 +1190,7 @@ class ValidationTest
                    .map(_.changeTimestamp(timeC))
         _      <- blockStorage.put(blockC.blockHash, blockC, Seq.empty)
         dag    <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task].deployHeaders(blockC, dag, chainName).attempt
+        result <- Validation.deployHeaders[Task](blockC, dag, chainName).attempt
       } yield result shouldBe Right(())
   }
 
@@ -1208,7 +1202,7 @@ class ValidationTest
         genesis <- createAndStoreBlock[Task](Seq.empty, deploys = deploysWithCost)
         block   <- createAndStoreBlock[Task](Seq(genesis.blockHash), deploys = deploysWithCost)
         dag     <- dagStorage.getRepresentation
-        result  <- ValidationImpl[Task].deployUniqueness(block, dag).attempt
+        result  <- Validation.deployUniqueness[Task](block, dag).attempt
       } yield result shouldBe Left(ValidateErrorWrapper(InvalidRepeatDeploy))
   }
 
@@ -1222,7 +1216,7 @@ class ValidationTest
                     deploys = deploysWithCost ++ deploysWithCost
                   )
         dag    <- dagStorage.getRepresentation
-        result <- ValidationImpl[Task].deployUniqueness(genesis, dag).attempt
+        result <- Validation.deployUniqueness[Task](genesis, dag).attempt
       } yield result shouldBe Left(ValidateErrorWrapper(InvalidRepeatDeploy))
   }
 
@@ -1321,7 +1315,7 @@ class ValidationTest
         c       <- createValidatorBlock[Task](Seq(genesis), bonds, Seq(a), v1)
         d       <- createValidatorBlock[Task](Seq(b), bonds, Seq(c), v0)
         dag     <- dagStorage.getRepresentation
-        _ <- ValidationImpl[Task].swimlane(d, dag).attempt shouldBeF Left(
+        _ <- Validation.swimlane[Task](d, dag).attempt shouldBeF Left(
               ValidateErrorWrapper(SwimlaneMerged)
             )
       } yield ()
@@ -1346,7 +1340,7 @@ class ValidationTest
         c       <- createValidatorBlock[Task](Seq(genesis), bonds, Seq(a), v1)
         d       <- createValidatorBlock[Task](Seq(a), bonds, Seq(c), v0)
         dag     <- dagStorage.getRepresentation
-        _       <- ValidationImpl[Task].swimlane(d, dag).attempt shouldBeF Right(())
+        _       <- Validation.swimlane[Task](d, dag).attempt shouldBeF Right(())
       } yield ()
   }
 
@@ -1367,12 +1361,12 @@ class ValidationTest
                    messageType = MessageType.BALLOT,
                    chainName = chainName
                  )
-        _ <- ValidationImpl[Task]
+        _ <- Validation[Task]
               .blockSummary(BlockSummary.fromBlock(blockA), chainName)
               .attempt shouldBeF Left(
               ValidateErrorWrapper(InvalidTargetHash)
             )
-        _ <- ValidationImpl[Task]
+        _ <- Validation[Task]
               .blockSummary(BlockSummary.fromBlock(blockB), chainName)
               .attempt shouldBeF Left(
               ValidateErrorWrapper(InvalidTargetHash)
