+++ client/src/main/scala/io/casperlabs/client/DeployRuntime.scala
@@ -19,6 +19,7 @@ import io.casperlabs.crypto.codec.{Base16, Base64}
 import io.casperlabs.crypto.hash.Blake2b256
 import io.casperlabs.crypto.signatures.SignatureAlgorithm.Ed25519
 import io.casperlabs.shared.FilesAPI
+import io.casperlabs.models.DeployImplicits._
 import org.apache.commons.io._
 import scalapb_circe.JsonFormat
 
@@ -504,32 +505,8 @@ object DeployRuntime {
   private def processError(t: Throwable): Throwable =
     Option(t.getCause).getOrElse(t)
 
-  private def hash[T <: scalapb.GeneratedMessage](data: T): ByteString =
-    ByteString.copyFrom(Blake2b256.hash(data.toByteArray))
-
   def readFileAsString[F[_]: Sync](file: File): F[String] = Sync[F].delay {
     new String(Files.readAllBytes(file.toPath), StandardCharsets.UTF_8)
   }
 
-  implicit val ordering: Ordering[ByteString] =
-    Ordering.by[ByteString, String](bs => Base16.encode(bs.toByteArray))
-
-  implicit class DeployOps(d: consensus.Deploy) {
-    def withHashes = {
-      val h = d.getHeader.withBodyHash(hash(d.getBody))
-      d.withHeader(h).withDeployHash(hash(h))
-    }
-
-    def sign(privateKey: PrivateKey, publicKey: PublicKey): Deploy = {
-      val sig = Ed25519.sign(d.deployHash.toByteArray, privateKey)
-      val approval = consensus
-        .Approval()
-        .withApproverPublicKey(ByteString.copyFrom(publicKey))
-        .withSignature(
-          consensus.Signature().withSigAlgorithm(Ed25519.name).withSig(ByteString.copyFrom(sig))
-        )
-      val approvals = d.approvals.toList :+ approval
-      d.withApprovals(approvals.distinct.sortBy(_.approverPublicKey))
-    }
-  }
 }
